requirements.txt
annotated-types==0.7.0
anyio==4.5.0
asgiref==3.8.1
attrs==24.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.3
black==24.8.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==2.0.12
chromedriver-autoinstaller==0.6.4
click==8.1.7
colorama==0.4.6
distro==1.9.0
Django==4.2
django-cors-headers==4.4.0
django-seed==0.3.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.2
Faker==29.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
iamport==0.0.1
iamport-rest-client==0.9.0
idna==3.10
jiter==0.5.0
mypy-extensions==1.0.0
openai==1.50.2
outcome==1.3.0.post0
packaging==24.1
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.3.6
pycparser==2.22
pydantic==2.9.2
pydantic-core==2.23.4
PyJWT==2.9.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
requests==2.32.3
selenium==4.25.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.1
tomli==2.0.1
toposort==1.10
tqdm==4.66.5
trio==0.26.2
trio-websocket==0.11.1
typing-extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
websocket-client==1.8.0
wsproto==1.2.0
